splitting learned abstraction in mole and rna
