+++ graphlearn/abstract_graphs/deep.py
